feat(ui): propagar teclado numérico a todos los diálogos getDouble (Opción A)
Generaliza el diálogo del teclado numérico y lo aplica a los 7 sitios de captura
numérica que usaban QInputDialog.getDouble crudo (violando SPJ_REFACTOR_SKILL
Regla 22/23 por su default forzado con min=0.001 o value pre-cargado).

Componente:
- NEW frontend/desktop/components/numeric_keypad_dialog.py — NumericKeypadDialog
  base genérico (cantidades, pesos, montos, %). Teclado grande táctil + toggle
  desplegable, inicio en 0/vacío, validación de rango/decimales, y flag
  permitir_cero para capturas donde 0 es válido (p.ej. fondo de caja inicial).
  Rechaza explícitamente valores > máximo (QDoubleValidator los trata como
  intermedios). API estática get_value() drop-in de QInputDialog.getDouble.
- quantity_input_dialog.py ahora es una subclase delgada (QuantityInputDialog +
  get_quantity) — ventas y sus tests siguen igual.

Sitios migrados:
- ventas.py: descuento % (decimals=1, max=100, unidad %).
- caja.py: fondo de apertura de turno (permitir_cero, unidad $).
- produccion.py: peso real de salida por componente (inicial=delta, permitir_cero, kg).
- finanzas_unificadas.py: cobro global CxC y pago global CxP (montos $).
- fidelidad_config.py: reservar presupuesto de rifa y costo real de premio (premio permite 0).

Guardrail: test_sales_manual_quantity_defaults_zero se actualiza para exigir el
componente estándar (inicio en 0/vacío) en vez del viejo patrón getDouble(value=0.0).

Tests: +test_numeric_keypad_dialog (5) — permitir_cero, tope máximo, valor inicial,
subclase. Total keypad 15 en verde. Suite de arquitectura 35/224 idéntica al baseline.
