+++ pos_spj_v13.4/modulos/fidelidad_config.py
@@ -613,7 +613,8 @@ class ModuloFidelidadConfig(QWidget):
     def _on_reservar_presupuesto(self):
         row = self._require_selected_raffle()
         if not row: return
-        monto, ok = QInputDialog.getDouble(self, "Reservar presupuesto", "Monto:", 0.0, 0.0, 99999999.0, 2)
+        from frontend.desktop.components.numeric_keypad_dialog import NumericKeypadDialog
+        monto, ok = NumericKeypadDialog.get_value(self, "Reservar presupuesto", "Monto:", decimals=2, maximo=99999999.0, unidad="$")
         if not ok or monto <= 0: return
         try:
             self.container.loyalty_service.reserve_raffle_budget(int(row["id"]), float(monto), self.usuario or "sistema", f"ui:reserve:{row['id']}")
@@ -664,7 +665,8 @@ class ModuloFidelidadConfig(QWidget):
         winner_id, ok = QInputDialog.getText(self, "Entregar premio", "ID del ganador:", text=suggested)
         winner_id = str(winner_id or "").strip()
         if not ok or not winner_id: return
-        costo, ok2 = QInputDialog.getDouble(self, "Entregar premio", "Costo real:", 0.0, 0.0, 99999999.0, 2)
+        from frontend.desktop.components.numeric_keypad_dialog import NumericKeypadDialog
+        costo, ok2 = NumericKeypadDialog.get_value(self, "Entregar premio", "Costo real:", decimals=2, maximo=99999999.0, unidad="$", permitir_cero=True)
         if not ok2: return
         try:
             self.container.loyalty_service.mark_prize_delivered(winner_id, self.usuario or "sistema", float(costo), f"ui:winner:{winner_id}")
